requirements.txt
streamlit==1.52.2
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.1
scikit-learn==1.3.0
imageio==2.37.3